chore: extract StaticCommand to own file
